updated renderer + added setupEnv package
the setupEnv package will contain all the functionality for setting up the maze and spawning in units at the start of a simulation. IMO much cleaner to have in a separate package that cluttering up main.

In setupEnv I am currently adding in code which spawns the walls to produce a maze